requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.12
email-validator==2.2.0

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
aiomysql==0.2.0
cryptography==44.0.0
redis==5.0.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
